Add --identifier for application

+++ jobbergate_cli/jobbergate_api_wrapper.py
@@ -507,7 +507,8 @@ class JobbergateApi:
             supplied_params = {}
 
         app_data = self.jobbergate_request(
-            method="GET", endpoint=urljoin(self.api_endpoint, f"/application/{application_id}")
+            method="GET",
+            endpoint=urljoin(self.api_endpoint, f"/application/{application_id}"),
         )
         if "error" in app_data.keys():
             return app_data
@@ -649,7 +650,8 @@ class JobbergateApi:
             return response
 
         response = self.jobbergate_request(
-            method="GET", endpoint=urljoin(self.api_endpoint, f"/job-script/{job_script_id}")
+            method="GET",
+            endpoint=urljoin(self.api_endpoint, f"/job-script/{job_script_id}"),
         )
         if "error" in response.keys():
             return response
@@ -690,7 +692,8 @@ class JobbergateApi:
             return response
 
         data = self.jobbergate_request(
-            method="GET", endpoint=urljoin(self.api_endpoint, f"/job-script/{job_script_id}")
+            method="GET",
+            endpoint=urljoin(self.api_endpoint, f"/job-script/{job_script_id}"),
         )
         if "error" in data.keys():
             return data
@@ -719,7 +722,8 @@ class JobbergateApi:
             return response
 
         response = self.jobbergate_request(
-            method="DELETE", endpoint=urljoin(self.api_endpoint, f"/job-script/{job_script_id}")
+            method="DELETE",
+            endpoint=urljoin(self.api_endpoint, f"/job-script/{job_script_id}"),
         )
 
         return response
@@ -779,7 +783,8 @@ class JobbergateApi:
         data["job_submission_owner"] = self.user_id
 
         job_script = self.jobbergate_request(
-            method="GET", endpoint=urljoin(self.api_endpoint, f"/job-script/{job_script_id}")
+            method="GET",
+            endpoint=urljoin(self.api_endpoint, f"/job-script/{job_script_id}"),
         )
         if "error" in job_script.keys():
             return job_script
@@ -787,7 +792,8 @@ class JobbergateApi:
         application_id = job_script["application"]
 
         application = self.jobbergate_request(
-            method="GET", endpoint=urljoin(self.api_endpoint, f"/application/{application_id}")
+            method="GET",
+            endpoint=urljoin(self.api_endpoint, f"/application/{application_id}"),
         )
         if "error" in application.keys():
             return application
@@ -888,7 +894,9 @@ class JobbergateApi:
         # TODO how to collect data that will updated for the job-submission
         response = self.jobbergate_request(
             method="PUT",
-            endpoint=urljoin(self.api_endpoint, f"/job-submission/{job_submission_id}/"),
+            endpoint=urljoin(
+                self.api_endpoint, f"/job-submission/{job_submission_id}/"
+            ),
         )
         return response
 
@@ -946,7 +954,13 @@ class JobbergateApi:
             return response
 
     @tabulate_decorator
-    def create_application(self, application_name, application_path, application_desc):
+    def create_application(
+        self,
+        application_name,
+        application_identifier,
+        application_path,
+        application_desc,
+    ):
         """
         CREATE an application.
 
@@ -968,6 +982,14 @@ class JobbergateApi:
                 solution="Please try again with --application-path specified",
             )
             parameter_check.append(response)
+
+        if application_identifier is None:
+            response = self.error_handle(
+                error="--identifier not defined",
+                solution="Please try again with --identifier specified",
+            )
+            parameter_check.append(response)
+
         if len(parameter_check) > 0:
             response = parameter_check
             return response
@@ -981,6 +1003,7 @@ class JobbergateApi:
         data = self.application_config
         data["application_name"] = application_name
         data["application_owner"] = self.user_id
+        data["application_identifier"] = application_identifier
 
         if application_desc:
             data["application_description"] = application_desc
@@ -1011,35 +1034,65 @@ class JobbergateApi:
         return response
 
     @tabulate_decorator
-    def get_application(self, application_id):
+    def get_application(self, application_id, application_identifier):
         """
         GET an Application.
 
         Keyword Arguments:
-            application_id -- id of application to be returned
+            application_id         -- id of application to be returned
+            application_identifier -- id of application to be returned
         """
-        response = self.jobbergate_request(
-            method="GET", endpoint=urljoin(self.api_endpoint, f"/application/{application_id}")
-        )
+        parameter_check = []
+        if application_id and application_identifier:
+            response = self.error_handle(
+                error="both identifier and id supplied",
+                solution="Please try again with only one",
+            )
+            parameter_check.append(response)
+
+        if len(parameter_check) > 0:
+            response = parameter_check
+            return response
+        if application_id:
+            response = self.jobbergate_request(
+                method="GET",
+                endpoint=urljoin(self.api_endpoint, f"/application/{application_id}"),
+            )
+        else:
+            response = self.jobbergate_request(
+                method="GET",
+                endpoint=urljoin(
+                    self.api_endpoint,
+                    f"/application/?identifier={application_identifier}",
+                ),
+            )
 
         return response
 
     @tabulate_decorator
-    def update_application(self, application_id, application_path, application_desc):
+    def update_application(self, application_id, application_identifier, application_path, application_desc):
         """
         UPDATE an Application.
 
         Keyword Arguments:
-            application_id    -- id application to update
-            application_path  --  path to dir for updated application files
-            application_desc  --  optional new application description
+            application_id            -- id application to update
+            application_identifier    -- identifier application to update
+            application_path          --  path to dir for updated application files
+            application_desc          --  optional new application description
         """
+        parameter_check = []
         if application_path is None:
             response = self.error_handle(
                 error="--application-path not defined",
                 solution="Please try again with --application-path specified",
             )
-            return response
+            parameter_check.append(response)
+        if application_id and application_identifier:
+            response = self.error_handle(
+                error="both identifier and id supplied",
+                solution="Please try again with only one",
+            )
+            parameter_check.append(response)
 
         error_check = self.application_error_check(application_path)
 
@@ -1047,9 +1100,17 @@ class JobbergateApi:
             response = error_check
             return response
 
-        data = self.jobbergate_request(
-            method="GET", endpoint=urljoin(self.api_endpoint, f"/application/{application_id}")
-        )
+        if id:
+            data = self.jobbergate_request(
+                method="GET",
+                endpoint=urljoin(self.api_endpoint, f"/application/{application_id}"),
+            )
+        else:
+            data = self.jobbergate_request(
+                method="GET",
+                endpoint=urljoin(self.api_endpoint, f"/application/?identifier={application_identifier}"),
+            )
+
         if "error" in data.keys():
             return data
 
@@ -1084,16 +1145,38 @@ class JobbergateApi:
         return response
 
     @tabulate_decorator
-    def delete_application(self, application_id):
+    def delete_application(self, application_id, application_identifier):
         """
         DELETE an Application.
 
         Keyword Arguments:
-            application_id -- id of application to delete
+            application_id         -- id of application to delete
+            application_identifier -- identifier of application to delete
         """
-        response = self.jobbergate_request(
-            method="DELETE",
-            endpoint=urljoin(self.api_endpoint, f"/application/{application_id}"),
-        )
+        parameter_check = []
+        if application_id and application_identifier:
+            response = self.error_handle(
+                error="both identifier and id supplied",
+                solution="Please try again with only one",
+            )
+            parameter_check.append(response)
+
+        if len(parameter_check) > 0:
+            response = parameter_check
+            return response
+
+        if application_id:
+            response = self.jobbergate_request(
+                method="DELETE",
+                endpoint=urljoin(self.api_endpoint, f"/application/{application_id}"),
+            )
+        else:
+            response = self.jobbergate_request(
+                method="DELETE",
+                endpoint=urljoin(
+                    self.api_endpoint,
+                    f"/application/?identifier={application_identifier}",
+                ),
+            )
 
         return response
